+++ EssentialFeed/Feed API/FeedItemsMapper.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-final class FeedItemsMapper {
+public final class FeedItemsMapper {
     
     private struct Root: Decodable {
         private let items: [RemoteFeedItem]
@@ -24,7 +24,7 @@ final class FeedItemsMapper {
         }
     }
     
-    static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
+    public static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
         guard response.isOK, let root = try? JSONDecoder().decode(Root.self, from: data) else {
             throw RemoteFeedLoader.Error.invalidData
         }
